CODE RUB: Adjusting namespaces and separating concerns with folders
