--- Standard.AI.Data.EntityIntelligence/Models/Datas/Services/ColumnMetadata.cs
@@ -1,12 +1,0 @@
-﻿// ---------------------------------------------------------------------------------- 
-// Copyright (c) The Standard Organization, a coalition of the Good-Hearted Engineers 
-// ----------------------------------------------------------------------------------
-
-namespace Standard.AI.Data.EntityIntelligence.Models.Datas.Services
-{
-    public class ColumnMetadata
-    {
-        public string Name { get; init; }
-        public string DataType { get; init; }
-    }
-}
