+++ Standard.AI.Data.EntityIntelligence/Models/Datas/Services/Informations/ColumnMetadata.cs
@@ -2,14 +2,11 @@
 // Copyright (c) The Standard Organization, a coalition of the Good-Hearted Engineers 
 // ----------------------------------------------------------------------------------
 
-using System.Collections.Generic;
-
-namespace Standard.AI.Data.EntityIntelligence.Models.Datas.Services
+namespace Standard.AI.Data.EntityIntelligence.Models.Datas.Services.Informations
 {
-    public class TableMetadata
+    internal class ColumnMetadata
     {
-        public string Schema { get; init; }
         public string Name { get; init; }
-        public IEnumerable<ColumnMetadata> ColumnsMetadata { get; init; }
+        public string DataType { get; init; }
     }
 }
