+++ Standard.AI.Data.EntityIntelligence/Models/Foundations/Datas/Exceptions/Informations/DataDependencyValidationException.cs
@@ -4,7 +4,7 @@
 
 using Xeptions;
 
-namespace Standard.AI.Data.EntityIntelligence.Models.Foundations.Datas.Exceptions
+namespace Standard.AI.Data.EntityIntelligence.Models.Foundations.Datas.Exceptions.Informations
 {
     internal class DataDependencyValidationException : Xeption
     {
