+++ Standard.AI.Data.EntityIntelligence/Services/Foundations/Datas/Informations/DataInformationService.Exceptions.cs
@@ -6,12 +6,13 @@ using System;
 using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Threading.Tasks;
-using Standard.AI.Data.EntityIntelligence.Models.Datas.Services;
-using Standard.AI.Data.EntityIntelligence.Models.Foundations.Datas.Exceptions;
+using Standard.AI.Data.EntityIntelligence.Models.Datas.Services.Informations;
+using Standard.AI.Data.EntityIntelligence.Models.Foundations.Datas.Exceptions.Informations;
+using Standard.AI.Data.EntityIntelligence.Services.Foundations.Datas.Informations;
 
 namespace Standard.AI.Data.EntityIntelligence.Services.Foundations.Datas
 {
-    internal partial class DataService : IDataService
+    internal partial class DataInformationService : IDataInformationService
     {
         private delegate ValueTask<IEnumerable<TableMetadata>> ReturningTableMetadatasFunction();
